+++ backend/src/main/java/net/es/oscars/nsi/svc/NsiService.java
@@ -18,6 +18,7 @@ import net.es.nsi.lib.soap.gen.nsi_2_0.services.point2point.P2PServiceBaseType;
 import net.es.nsi.lib.soap.gen.nsi_2_0.services.types.DirectionalityType;
 import net.es.nsi.lib.soap.gen.nsi_2_0.services.types.OrderedStpType;
 import net.es.nsi.lib.soap.gen.nsi_2_0.services.types.StpListType;
+import net.es.nsi.lib.soap.gen.nsi_2_0.services.types.TypeValueType;
 import net.es.oscars.app.exc.NsiException;
 import net.es.oscars.app.exc.PCEException;
 import net.es.oscars.dto.pss.cmd.CommandType;
@@ -154,8 +155,6 @@ public class NsiService {
                     }
 
                 } else {
-                    // delete the mapping as this failed
-                    nsiRepo.delete(mapping);
                     log.error("error reserving");
                     nsiStateEngine.reserve(NsiEvent.RESV_FL, mapping);
                     try {
@@ -164,6 +163,8 @@ public class NsiService {
                                 result.getErrorCode().toString(),
                                 result.getTvps(),
                                 header.getCorrelationId());
+                        nsiRepo.delete(mapping);
+
                     } catch (WebServiceException | ServiceException cex) {
                         log.error("reserve failed: then callback failed", cex);
                     }
@@ -171,13 +172,14 @@ public class NsiService {
             } catch (Exception ex) {
                 log.error("Internal error: " + ex.getMessage(), ex);
                 try {
-                    nsiRepo.delete(mapping);
                     nsiStateEngine.reserve(NsiEvent.RESV_FL, mapping);
                     this.errCallback(NsiEvent.RESV_FL, mapping,
                             "Internal error",
                             NsiErrors.NRM_ERROR.toString(),
                             new ArrayList<>(),
                             header.getCorrelationId());
+                    nsiRepo.delete(mapping);
+
                 } catch (Exception cex) {
                     log.error("reserve failed: then callback failed", cex);
                 }
@@ -647,9 +649,6 @@ public class NsiService {
                 qrrt.setResultId(resultId);
                 qrct.getReservation().add(qrrt);
                 resultId++;
-            } else {
-//                log.info("will delete an invalid nsi mapping for " + mapping.getNsiConnectionId() + " - " + mapping.getOscarsConnectionId());
-//                invalidMappings.add(mapping);
             }
 
         }
@@ -695,9 +694,6 @@ public class NsiService {
                 qsrt.setResultId(resultId);
                 qsct.getReservation().add(qsrt);
                 resultId++;
-//            } else {
-//                log.info("will delete an invalid nsi mapping for " + mapping.getNsiConnectionId() + " - " + mapping.getOscarsConnectionId());
-//                invalidMappings.add(mapping);
             }
         }
         nsiRepo.deleteAll(invalidMappings);
@@ -1359,6 +1355,11 @@ public class NsiService {
     public P2PServiceBaseType makeP2P(Components cmp, NsiMapping mapping) {
 
         P2PServiceBaseType p2p = new P2PServiceBaseType();
+
+        TypeValueType tvt = new TypeValueType();
+        tvt.setType("oscarsId");
+        tvt.setValue(mapping.getOscarsConnectionId());
+        p2p.getParameter().add(tvt);
 
         VlanFixture a = cmp.getFixtures().get(0);
         String srcStp = this.nsiUrnFromInternal(a.getPortUrn()) + "?vlan=" + a.getVlan().getVlanId();